Update TwigTemplateTask for upcoming bake changes.
To solve cakephp/bake#395 I had to add an additional parameter to the
public `bake` method. While this is OK from a backwards compatibility
point of view, it does trigger a strict error in this plugin. I've also
fixed the issue present in cakephp/bake#395 that was also found here.